[ADD]Folder 재구축 + Fetch 함수 추가 -하나의 repository에 합치기 위하여 폴더 재구축 -backend와의 통신을 위하여 fetch함수 추가
