+++ src/Routes.js
@@ -4,9 +4,9 @@ import {
   Route,
   Switch,
 } from 'react-router-dom';
-import Main from './pages/Main/Main';
-import Login from './pages/Login/Login';
-import SignUp from './pages/SignUp/SignUp';
+import Main from './pages/MainM/Main';
+import Login from './pages/LoginM/Login';
+import SignUp from './pages/SignUpM/SignUp';
 
 class Routes extends React.Component {
   render() {
