+++ src/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import "./styles/reset.scss";
+import "./stylesM/reset.scss";
 import Routes from './Routes';
 ReactDOM.render(<Routes />,  document.getElementById('root'));
