+++ src/pages/MainM/components/RightInfo/RightInfo.scss
