+++ src/pages/MainM/components/RightStories/RightStories.scss
