+++ src/pages/MainM/components/RightSuggestions/RightSuggestions.scss
