+++ src/pages/MainM/components/Story/Story.scss
