+++ src/pages/MainM/components/Suggestion/Suggestion.scss
